--- client/internal/text/serializer/custom.ts
@@ -1,82 +1,0 @@
-import React from "react";
-import { RichElement, ISerializationRegistryType } from ".";
-import { CUSTOM_CLASS_PREFIX } from "..";
-import { serializeElementBase, deserializeElementBase, deserializeElement, IElementBase, reactifyElementBase, IReactifyTemplateOptions } from "./base";
-
-export function registerCustom(registry: ISerializationRegistryType) {
-  const boundDeserializeElement = deserializeElement.bind(null, registry);
-
-  function serializeCustom(custom: ICustom) {
-    return serializeElementBase(
-      registry,
-      custom,
-      "div",
-      CUSTOM_CLASS_PREFIX + custom.customType,
-      null,
-      custom.children,
-    );
-  }
-  
-  function deserializeCustom(node: HTMLDivElement): ICustom {
-    const base = deserializeElementBase(node);
-    let customType: string = null;
-    node.classList.forEach((c) => {
-      if (c.startsWith(CUSTOM_CLASS_PREFIX)) {
-        customType = c.substr(CUSTOM_CLASS_PREFIX.length);
-      }
-    });
-    const custom: ICustom = {
-      ...base,
-      type: "custom",
-      containment: "superblock",
-      customType,
-      children: Array.from(node.childNodes).map(boundDeserializeElement).filter((n) => n !== null) as RichElement[],
-    }
-    return custom;
-  }
-
-  function reactifyCustom(
-    custom: ICustom,
-    active: boolean,
-    customProps?: React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement>,
-    templateOptions?: IReactifyTemplateOptions,
-  ) {
-    return reactifyElementBase(
-      registry,
-      active,
-      custom,
-      "div",
-      CUSTOM_CLASS_PREFIX + custom.customType,
-      null,
-      customProps,
-      custom.children,
-      templateOptions,
-    );
-  }
-
-  registry.REACTIFY.custom = reactifyCustom;
-  registry.SERIALIZE.custom = serializeCustom;
-  registry.DESERIALIZE.byClassNamePrefix.custom = deserializeCustom;
-}
-
-/**
- * The custom type represents a custom- element
- */
-export interface ICustom extends IElementBase {
-  /**
-   * The type as custom
-   */
-  type: "custom";
-  /**
-   * refers to be able to contain blocks or other super blocks, etc...
-   */
-  containment: "superblock",
-  /**
-   * Specifies which custom type it is
-   */
-  customType: string;
-  /**
-   * The children
-   */
-  children: RichElement[];
-}
